pyproject.toml
[project]
name = "fsion"
version = "0.1.0"
description = """FSion (fusion: /ˈfjuʒən/) is a library designed to expose
filesystem-like API's from various backends (Google Drive, S3, GCS, local
filesystem etc.)
"""
readme = "README.md"
authors = [
    { name = "Roni Bytyqi", email = "imscytha@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.1",
]

[project.scripts]
fsion = "fsion:main"

[project.optional-dependencies]
drive = [
    "google-api-python-client>=2.166.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
